[Event Hubs] Improve tests
