+++ sdk/eventhub/event-hubs/test/public/utils/EnvVarKeys.ts
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export enum EnvVarKeys {
+  EVENTHUB_CONNECTION_STRING = "EVENTHUB_CONNECTION_STRING",
+  EVENTHUB_NAME = "EVENTHUB_NAME",
+  AZURE_TENANT_ID = "AZURE_TENANT_ID",
+  AZURE_CLIENT_ID = "AZURE_CLIENT_ID",
+  AZURE_CLIENT_SECRET = "AZURE_CLIENT_SECRET",
+  TEST_TARGET = "TEST_TARGET",
+}
