+++ sdk/eventhub/event-hubs/test/public/utils/getEnvVarValue.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import "./env";
+
+export function getEnvVarValue(name: string): string | undefined {
+  return process.env[name];
+}
